fixed users.test

+++ tests/integration/users.test.js
@@ -9,7 +9,7 @@ const root = '/users';
 describe(root, () => {
   beforeEach(() => { 
     server = require('../../index'); 
-    token = new User().generateAuthToken();
+    // token = new User().generateAuthToken();
   })
   afterEach(async () => { 
     await server.close(); 
@@ -26,14 +26,24 @@ describe(root, () => {
       const token = user.generateAuthToken();
       const decoded = jwt.verify(token, config.get('jwtPrivateKey'));
       const revert = await User.deleteOne({ name: 'jwt' });
-      expect(decoded).toMatchObject(payload);
+      // console.log('user func: ', user);
+      // console.log('token: ', token);
+      // console.log('decoded: ', decoded);
+      // expect(decoded).toMatchObject(payload);
+      // expect(decoded._id).toBe(payload._id);
+      expect(decoded.isAdmin).toBe(payload.isAdmin);
     });
   });
 
-  describe('GET /me', () => {
+  describe('GET /me/:id', () => {
     it('should return current user successfully', async() => {
       const user = await User.findOne({ isAdmin: false });
-      const res =  await request(server).get(`${root}/me/${user._id}`)
+      const address = `${root}/me/${user._id}`;
+      const token = user.generateAuthToken();
+      const res =  await request(server).get(address).set('x-auth-token', token);
+      // console.log('address: ', address);
+      // console.log('get user: ', user);
+      // console.log('get res.body: ', res.body);
       expect(res.status).toBe(200);
       expect(res.body).toHaveProperty('name', 'normal');
     })
@@ -41,24 +51,44 @@ describe(root, () => {
 
   describe('POST /', () => {
     it('should return 400 if user already registered.', async() => {
-      const user = User.findOne({ isAdmin: false });
+      const user = {
+        name: 'normal',
+        email: 'normal@gmail.com',
+        password: 'normalpas',
+        isAdmin: false   
+      }
       const res = await request(server).post(root).send(user);
-      console.log('user: ', user);
-      console.log('res.body: ', res.body);
+      console.log('400 post user: ', user);
+      console.log('400 post res.body: ', res.body);
       expect(res.status).toBe(400);
     })
     it('should return 200 if everything is good', async() => {
       const user = {
+        _id: new mongoose.Types.ObjectIdThe 12-byte ObjectId consists of:
+
+A 4-byte timestamp, representing the ObjectId's creation, measured in seconds since the Unix epoch.
+
+A 5-byte random value generated once per process. This random value is unique to the machine and process.
+
+A 3-byte incrementing counter, initialized to a random value.
+
+While the BSON format itself is little-endian, the timestamp and counter values are big-endian, the most significant bytes appear first in the byte sequence.
+
+If an integer value is used to create an ObjectId, the integer replaces the timestamp.().toHexString(), 
         name: 'temp',
         email: 'temp@gmail.com',
         password: 'temp123',
         isAdmin: false
       }
-      const token = await user.generateAuthToken();
-      const res = await request(server).post(root).send(user);
+      const res = await request(server).post(root).send(user);    
       const revert = await User.deleteOne({ name: 'temp' });
+      // console.log('200 post user: ', user);
+      // console.log('200 post res: ', res);
+      const token = res.headers['x-auth-token'];
+      const decoded = jwt.verify(token, config.get('jwtPrivateKey'));
       expect(res.status).toBe(200);
-      expect(res.header).toEqual(token);
+      expect(decoded.isAdmin).toBe(false);
+      // expect(decoded._id).toBe(user._id);
       expect(res.body).toHaveProperty('name', 'temp');
     })
   })
